Began work on angularified add prefix page
Added a Pylons controller to handle an AngularJS-version of the add
prefix page. Created a template, basically a copy of the original add
prefix page, where AngularJS model bindings will be added. Added an
(empty) angular controller which will handle the add prefix page as well
as added possible AngularJS-routes to the application configuration. As
the latter requires the angular-route module, this was included as well.