Trigger restructure for tag handling and indent calc
The before trigger for tags has been merged with our large before
trigger.

It now also handles DELETEs which is a prerequisite to correctly handle
inherited tags. The trigger is also set for a few more actions, like
when tags are changed. This means that we might do more computations
more often.. not sure of performance impact.

Now that we have a larger trigger function shared by lot of different
stuff, we need to think about when the function is triggered and not act
on all updates. This adds a check for indent calc so that the prefix
actually has been update before we run calc indent.

Part of #248.