Auto-prompt for recursive delete of assignment
If a user tries to delete an assignment that contains hosts without the
recursive option it will not be allowed, so we prompt the user if he
wants to do a recursive delete if the assignment contains hosts.

Since the recursive option is a bit "hidden" this greatly enhances
usability.

Fixes #520.