Add basic pool statistics
Triggers in the database will update total number of addreses in a pool.

CLI includes statistics on total number of addresses in a pool as well
as number of used and free addresses.

There's a new test suite to check that used / free addresses are updated
when adding new child prefixes, which is essentially when the parent
prefix is part of a pool as well as updating the total, free and number
of child prefixes when adding and removing members to a pool. We also
mangle pool results in xmlrpc.py since we don't care about return values
there - all statistics testing is done in nipaptest.py

Includes the fugly hack to send large numbers as strings due to lack of
support for large numbers in the XML-RPC standard. Pynipap converts back
to numbers on the client side.

Part of #248.