Add columns for total / free prefixes in pools
We already have counters for keeping track of how many used/free/total
addresses we have in a pool but in the majority of cases pools are used
to assign larger prefixes and not individual addresses so it is more
interesting to know the number of free prefixes (based on default
assignment size). These new columns are for precisely that purpose.

The total number of prefixes (used + free) is an interesting value as it
will change depending on the size of the prefixes in the pool.

Part of #248.