Add option for disabling pynipap object cache
Pynipap caches objects per default to lower the amount of requests for
objects that occur a lot, typically VRFs. This can cause problems
sometimes, especially as there is no expiration of old entries in the
cache.

To temporarily solve some problems, this implements the option to
disable the cache. The cache is still kept up to date for deletes,
updates and so forth, the only difference is that we don't use it when
getting objects. This potentially creates some flexibility in using the
cache selectively (ie, disabling temporarily)