Add new test target to verify NIPAP upgrade
First install stable version of NIPAP from official debian repo and then
build new version from source to test the upgrade process of NIPAP in
addition to just running the more static tests.

After the first version from the official debian repo has been
installed, the script tests/upgrade-before.py is run to insert some data
into the database. NIPAP is then upgraded to the version built from
source after which tests/upgrade-after.py is run to verify the output.
upgrade-after.py needs to be updated upon database schema changes to
make sure that the correct columns are there as well as the correct
values are set after the upgrade.

To guarantee that an upgrade is always performed, the version number is
faked to 9999.9.9

Fixes #524.