Add statistics test to upgrade test suite
The recently introduced upgrade test suite allows the testing of data
over the NIPAP upgrade process from the stable version to the latest git
version. Since we have introduced SQL schema changes with the addition
of statistics, we use this test suite to make sure that data entered
before the upgrade comes out correct after the upgrade. That includes
having the new statistics columns / attributes as well as having the
correct values! :)

Part of #248.