Add option disable interactive prompting
Certain actions, typically removing prefixes, pools or VRFs results in
an interactive prompt to ensure that the user hasn't mistyped anything
and is really sure he wants to perform the action.

There are situations, commonly when scripting, where it would be
beneficial to disable the prompting and instead assume that the user
knows what he is doing. This change introduces the --force option which
allows the user to suppress any interactive prompting.

Fixes #499.