Fix Chart.JS setup for pools
The canvas is only included if pool has members of given address family
but we always tried to draw to the canvas (which didn't exist if there
were no members). This resulted in a JavaScript error and if the pool
contained IPv6 prefixes but no IPv4 prefixes, the chart for IPv6
wouldn't be drawn. This change fixes that.

I would like to move all of this to JavaScript on the client side
instead of mixing up server side Jinja2 with Javascript, but that will
probably come once we migrate these pages to AngularJS.

Fixes #552.