Fix ambiguous return value for nipap-passwd --latest-version
Before, the function in authlib.py to check the sqlite database for
latest version simply returned an exception if the db was out of date.

In this approach, the function will return false and nipap-passwd will
exit with the code 2. This way nipapd.config does not catch random
exceptions (exit code 1) anymore.