Further improve CLI "colon" tab completion
This changes the way tab completion works with regards to colon.

Given we have VRFs 123:123 and 123:456. Before change would yield
$ nipap address add vrf_rt 123:<tab>
$ nipap address add vrf_rt 123:123\:
Which is clearly broken. However this works:
$ nipap address add vrf_rt 123\:<tab>
123:123 123:456

While after the change, we fix the first behaviour
$ nipap address add vrf_rt 123:<tab>
123 456
But the completion lists only the part right of the colon and not the
complete VRF RT, which could be perceived as incorrect.

Kind of part of #422.