Update contribution guidelines and license
The contribution guidelines have been cleaned up as well as parts of the
README file. Thanks to the Bootstrap team which has written such an
excellent CONTRIBUTING.md file that I just had to copy some (a lot) from
it. Documentation is licensed under creative commons. This wasn't very
clear before, but it is and now we have a docs/LICENSE file to prove it
;)

Also add the latest fixes and features to the NEWS file.