Disable Travis-CI pylint and S3 upload test
We want pylint results over time and the running of pylint together with
uploading the results to S3 was a step in that direction. Recently we
found Scrutinizer-CI which can achieve the same things and as the
Travis-CI build process is considerably lenghtened by the S3 upload, I
have now disabled that test. We will rely entirely on Scrutinizer-CI!

Part of #274.