SQL functions for improved pool statistics
A couple of new stored procedures and a new trigger has been added to
calculate the number of free and total prefixes available in a pool
based on the default prefix-length.

The idea behind this is to list the number of total / used / free
prefixes instead of just listing total / used / free addresses. Since
the default prefix-length of a pool is typically used when assigning new
prefixes from that pool, it becomes more interesting with the amount of
free prefixes than free addresses.

The number of used prefixes is defined as the number of direct childs to
a member prefix of the pool. Free prefixes should be the number of free
prefixes with the pool's default assignment size in all the member pool
while the total prefixes is used + free combined.

This leads to the slightly awkward side-effect that the total number of
prefixes in a pool can vary if non-default prefix-lengths are used.

Part of #528.