pynipap: support for improved pool statistics
Pynipap now also converts the new prefix statistics attributes on pools
similar to how its done for other statistics attributes.

One change is that the new columns might contain NULL values and if so,
we let them remain as NULL (None) to indicate that they could not be
computed.

Part of #528.