Handle upgrade case for fixed prefix statistics
Since previous versions contains buggy code for calculating the number
of used addresses in a prefix when the children were /32s of non-host
type, we need to recalculate all statistics to make sure it's correct
and we add a test to the upgrade test to make sure it works.

Fixes #590.