Fix stats calculation for non-host /32s
.. or /128s for IPv6. /32s or /128s (max prefix length) need to be
handled differently due to how the stats are calculated. Before this
change we looked if the type was set to host to determine if we were at
max prefix length, but since we can add a non-host (ie reservation or
assignment) as a /32 or /128, we would get incorrect statistics.

This changes the check to simply check for the prefix length instead of
the type of the prefix.