travis: add virtualenv build
This is a pretty invasive change to the Travis-CI configuration file.
Before this, we only ran our tests in an environment where nipap was
installed using apt packages (we build them first and then installed).
Now we add a second mode, which is running the whole nipap installation
and tests in a virtualenv.