Tightened up CSS selectors
The CSS selectors used to define the different blocks on many NIPAP
pages (classes option-groups and options) were very general and set
properties to every child element of the option divs. This for example
made the new, Angularified, VRF-selector menu look.. ugly.

To address this the CSS selectors were tightened up to only apply to
elements at specific places in the tree. I might have overlooked some of
the consequences of this...