Added Pylons ng-ctrllr for single-page-ish pages
Added a new 'ng'-controller to Pylons for single-page-ish AngularJS
pages. In the future this can probably be replaced with a static HTML
page, but currently I judged this the best balance between avoiding
touching too much code (for example the hilighting of the current
section in the top menu bar, which still is handled by the Jinja2
templates) and enabling AngularJS functionality.

Added a prefix-action to the ng-controller which handles prefix-related
pages.