Moved angularified add-prefix page to ng-ctrllr
Moved the AngularJS-version of the add-prefix page from an own action in
the Pylons prefix controller to the new Pylons 'ng'-controller. The
prefix allocation method (from-prefix, from-pool or manual) can be
passed to the controller as URL parameters, optionally together with a
parameter to the allocation method (ID of a prefix to allocate from in
case of from-prefix, ID of pool to allocate from in case of from-pool).

Also did some fixes for allocation from prefix to work properly.