backend: Semi-working improvement smart parsing
This shows how pyparsing could help us parse more complex search
queries. pyparsing translates it into an AST that is then parsed through
some code to convert it to dictSQL.

Proper grouping of operators isn't working, ie 'and' and 'or' have the
same priority which leads to incorrect queries. There are probably other
issues as well, but it's something :)