cli: add support for extra-attributes (AVPs)
AVPs are called just that - AVPs - through the database, backend and
pynipap code but in the CLI front-end we call them extra-attributes as
that was deemed somewhat more intuitive. It might be that they are even
more tightly integrated into the future, making them almost on par with
built-in attributes. For the backend however, we don't feel the same
need of hiding what AVPs really are and the concept should be familiar
to most programmers or users of the API.

The CLI uses the new multiple key in the command structure to allow the
input of multiple extra-attributes to the same prefix.

Two new commands are introduced;
 - nipap address modify X add extra-attribute foo=bar
 - nipap address modify X remove extra-attribute foo
To help better manipulate these new extra-attributes (AVPs).