travis: create hstore extension to fix test
The hstore extension is usually loaded from the ip_net.plsql file but as
the search_path is messed up on Travis-CI it seems unable to do that, so
instead we load the extension beforehand from the sql Makefile.

As far as I can see, this shouldn't be an issue for other virtualenv
environments (I can't reproduce it on my laptop). Not sure why it
happens on Travis-CI. Might have to revisit if other have difficulties
installing in virtualenv and figure out a way to modify search_path
instead.