www: Moved prefix form defs. to separate template
Moved the prefix form definition from the prefix add-template to a
separate template to prepare for re-use in an angularified prefix
edit-template. Prepared the controller & template by adding a variable
to the $scope to let the template know wherher it's used on an add- or
edit-page.

Refactored the code to keep state of whether the comment-field should be
visible or not.