cli: fix AVP support for prefix on from-pool
Previous work and PR for adding AVPs to prefixes missed out on adding it
to the from_pool operation, since that is split from the rest of the
add_prefix functionality. We really should have one parsing function for
it though to avoid these kinds of issues in the future.