www: Added AngularJS-version of prefix edit-page
The next step on our path towards porting the NIPAP web UI to AngularJS
has been taken, with the angulatification of the prefix edit-page.
Added a PrefixAddController, a template and routes to the new controller
& template.

Statistics needs to be added before the page can be considered in
feature parity with the old prefix edit-page.