www: Link to new add prefix page
Replaced the "Add prefix"-button on the prefix list page with an
AngularJS/Bootstrap dropdown menu with three options: Manual, From pool
and From prefix.

When choosing From pool or From prefix the user gets a message
explaining that the operations now are performed from the pool list and
prefix list pages respectively.