backend: add expires field
The expires field simply contains a date when the current assignment
expires to help an IP registry / administration team to keep track of
assignment, possibly removing them or renewing them when expired.

In the database the expires attribute is stored as a timestamp and since
PostgreSQL supports infinity, we use that to signify if there is no
expiration time for a prefix. Pythons datetime objects on the other hand
do not support infinity and instead we rely on None to signify
+infinity.

The basic format is thus None for infinity and a ISO8601 string for
setting a specific date, in addition there is also support (through
dateutil) to set times like "two years" which means current time + two
years.