backend: mangle new expires column
The new expires column uses infinity in the database (supported by
postgresql) but as that is not supported by Pythons datetime object we
cast infinity to None.

This is the second type of mangling we do, in addition to the existing
mangling of long numbers for the statistics. I've refactored the whole
thing a bit and moved out the mangling parts into a separate function.