www: Added pool select directive to prefix forms
Added a directive showing a popup where a pool can be selected. This was
added to the prefix add and edit forms, so now the pool a prefix belongs
to can be set directly on the prefix.

During the work with the directive a bug was discovered. The bug was
introduced when the prefix form definitions were moved to a separate
template included with ngInclude. ngInclude created a new scope, which
means that the $scope properties which are bound to the VRF and pool
selector directives are bound to ngInclude's scope and not the
controller's scope.

To fix this the VRF and pool variables were moved from the scope to the
prefix object.