backend: allow configuration of forks
There seems to be a bug that hits us on certain Debian machines where we
have concurrency issues overwriting data of the different worker
processes (don't ask me how that is possible).

It is related to the forking code and while it normally is a sane
default to fork it's simply not a good idea if it's not working :P

This code allows the configuration of number of forks. If the forks
option isn't set in the configuration file, we will default to not
forking.