Update debian control file for nipapd package
Add support for postgresql 9.4 (and ip4r).

Also changed the way we depend on postgresql, before we used to depend
on the debian package called "postgresql" which is a meta package that
points to a specific version such as "postgresql-9.1". The problem is if
the user wishes to install another version as our dependency on
"postgresql" would force which version is to be installed. It's not a
big problem - most of the time you want the version that "postgresql"
points to, but not always.