www-ui: Enable node-field for max prefix length
Re-implementation of functionality missed when rebuilding web UI with
AngularJS. The node field on the add/edit prefix-page should be enabled
when adding a prefix with max prefix length (/32 for IPv4 and /128 for
IPv6), no matter what prefix type has been selected.

I chose to place the code in a service which is injected into the prefix
add and prefix edit controllers.

Fixes #675