Refactor smart search parsing code and add test
The code for parsing a smart search prefix query wasn't very testable as
it issued a database query after parsing the query. This has been
refactored into two functions so that the parsing is done separately and
thus becomes testable.

Four tests are added to check for three different types of queries and a
fourth to check that we correctly detect non-balanced quotes.