Add query interpretation to CLI
This adds the possibility of displaying the query output with the CLI by
adding the --show-interpretation argument.

The output is copied from the web UI which means it might not be
perfectly suitable for a CLI - tooltips are used in the web UI to give a
deeper explanation while in the CLI the tooltip info is printed on a
second line. For certain cases the brief explanation and the more
verbose "tooltip" explanation are very similar and might seem redundant.

Fixes #682.