backend: Improve prefix query interpretation result
Instead of sending a separate data struct (which was just a list) with
the query interpretation, we modify the query dictSQL and insert our
interpretation into it which means that the change of discrepancy
between the actual query and the interpretation is minimized.

Naturally, this new format allows us to express a much more complex
query than before when we assumed all entries in the interpretation list
to be ANDed together.

This change only affects the smart_search_prefix function at this time
and subsequently the format for VRFs and pool searches remain the same.

Part of #683.