web-ui: accept new prefix interpretation format
The web UI now accepts the improved prefix query interpretation format
implemented in nipapd.

We need to work on the layout of the interpretation as it is quite
verbose and not so pretty in its current incarnation. I suppose we
should rewrite the whole thing using AngularJS as well, unfortunately
I'm on a plane and haven't worked enough with Angular to remember from
the top of my head how to implement this nicely.

Part of #683.