backend: fix error message for empty pool
Trying to assign addresses from an empty pool would give a weird error
message ("VRF must be the same as the pools implied VRF"). Since we run
find_free_prefix after we look if the requested VRF is equal to the
pools implied VRF I have introduced a check earlier on in the code to
make sure the implied VRF is set (which means there are prefixes in the
pool). The result is an error message that is much more relatable.

Fixes #685