cli: fix viewing of pool statistics
To calculate correct statistics there needs to be at least one member
prefix and the default prefix length need to be set for the address
family in question. When this is not the case, the CLI would throw a
traceback but by adding some checks (similar to the ones in the web UI)
we avoid division by None and instead display a message stating that
statistics aren't available and why.

I've added another issue (#693) to track the issue that the backend
calculates statistics differently for prefixes and addresses.

Fixes #672.