Add filter for requires.io
requires.io thinks we should bump the version dependency of psutil to
2.something but 1.x that we are using has a completely different
interface so we can't just upgrade without also changing our code.

Debian ships psutil 1.x which is why stay on that version. To stop
requires.io from opening up a PR to update to 2.x we add a filter to say
we want something below 2.0.