backend: late import of ldap module
This moves the import of the ldap module from the start of nipapd to
when it is actually needed. If LDAP isn't configured as backend
authentication then it will never be used.

If LDAP authentication is enabled but the ldap module is not available,
we will log an error message when someone tries to authenticate. It is
NOT logged on startup of nipapd which might not be completely obvious.
We would need to rework larger parts of the code to achieve that.

Fixes #701.