backend: make LdapAuth more robust
Instead of requiring ro_group and rw_group to be configured in the
configuration file, we simply assume them to be empty if not explicitly
configured.

For users that are not members of a group at all there will be no
memberOf value and so we avoid getting key errors by using .get()
instead.