whoisd: add IPv6 support
SocketServer is hard-coded to listen to IPv4 so this introduces some
monkey patching to get around that and enable it for listening on an
IPv6 socket. When binding to "" or :: the server will listen on both
IPv4 and IPv6. To listen on an IPv4 address it needs to be specified as
an IPv4-mapped IPv6 address.

Fixes #718.