Make pynipap python3 compatible
xmlrpclib is called xmlrpc.client in py3 so the import first checks
which version of Python we are running. I would prefer to write py3 code
which was compatible with py2, so for the import I'd prefer calling it
by its py3 name "xmlrpc.client", unfortunately that's not possible so
we reference it by xmlrpclib.

py3 doesn't have the concept of long, it is simply called int and
handles large numbers.

Credit for to dosomder for his work on #594. I'm kind of rewriting it
due to difficulty of merging that PR, so this closes #594.