www: fix expires field of infinity
There were two issues surrounding the handling of positive infinite
values for the expires field. Upon reading the values the JSON encoder
would always cast to string which doesn't work very well with None
values as that would be represented as the string "None" rather that
<nil> over the wire and as we use nil for presenting infinity, this
would mess things up. Properly handling this means the text input in the
web form stays empty.

Secondly, we mangle the expires field to format the field into a ISO8601
value before sending to the backend when editing/adding a prefix. We
shouldn't mangle if the expires field is empty, since this means
infinity and mangling it will mess it up.

Fixes #721.