backend: fix so last_modified is always updated
The before trigger that sets last_modified to NOW() had conditions so
that it would only be triggered when changes to vrf_id, type or prefix
happened but as the last_modified value needs to be touched for all
updates this wouldn't give the desired results.

Instead the trigger is now always called but there is a shortcut out of
it to reduce latency for most cases, i.e. when vrf_id, type or prefix
stays the same. The conditional trigger is essentially emulated by a
condition inside the trigger function.

I don't know why we were checking that the old and new pool_id were
equal. The check wasn't even a proper one as usually one of the values
would be NULL so it would return false despite both values being NULL.
I have removed this check - hope nothing breaks, but I can't see what
would.

last_modified should now always be updated!

Fixes #749.