cli: improve query interpretation output
Instead of printing AND operators on a separate line they are now on
the same line as a match statement making the output a lot more compact
in number of lines, although a bit wider. Through some ASCII art we try
to show which statements are joined together by a boolean operator.

I believe readability is greatly enhanced by this change, though that
might be biased ;)