cli: accept natural mask for prefix view
Previously the view prefix function wouldn't accept the display_prefix,
i.e. a prefix with its natural mask. To show a host prefix it had to be
entered either with a host mask, e.g. /32 or /128 or with no mask at all.

There are scenarios when a host has the network address, most typically
when using /31 or /127 networks. To distinguish between the two, the
mask must be entered, i.e. use /31 to show the network or /32 to show
the host.

Fixes #738.