backend: enable case-insensitive tag searches
PostgreSQL doesn't have very powerful array operators so doing a
case-insensitive search is actually kind of difficult by normal means
(like lower()). Instead I use the citext module which is a standard
extension, shipped with PostgreSQL since 8.x, that allows
case-insensitive matches. This in turn makes it easy to recast the tags
column as citext and match on.

Fixes #743.