backend: make case-insensitive IN ANY default
Matching on tags uses the = ANY operator since the tags are stored as an
array. I instinctively created a new operator that we call equals_any_ci
which would be the case-insensitive counter part of equals_any but as
our default is to mostly do case-insensitive matching I'm also changing
the meaning of equals_any to be case-insensitive. If we want we can add
a _cs equivalent later on that does case-sensitive matching.

Part of #743.