authlib: Separate LDAP search credentials
Added functionality to allow NIPAP administrators to specify separate
LDAP credentials used when performing the LDAP search for the user
logging in. This is needed for LDAP setups where ordinary user accounts
are not permitted to search the LDAP database.

The search for the user logging in is performed after the user has been
authenticated to get full name and permissions (if used).

Fixes #794