utilities: Script to bulk string replace prefixes
Added a script which can be used to do bulk string replacement on
prefixes in NIPAP. That is change description and node in bulk. When
executed, the script will display a list of the prefixes to change and
what they will be changed to as well as let the user select what
prefixes to perform the operation on. Either all, a few selected
prefixes or all prefixes EXCEPT the selected prefixes can be chosen.