cli: fix prefix type guessing
The logic to automatically guess type of a new prefix failed when trying
to add a prefix larger than one already existing but with the same
network mask as the search to find the parent prefix would ignore the
prefix length. This is now fixed!

For example, 10.0.0.0/24 is an assignment and we are trying to add
10.0.0.0/8 as a reservation. The auto type guess would strip away the
prefix length, resulting in a search for "10.0.0.0" which the backend
interprets as 10.0.0.0/32 and that returns 10.0.0.0/24 as the parent
prefix. Since 10.0.0.0/24 is an assignment, the type would be changed to
"host" and the new prefix would be created with a /32 mask.

Now we check if the user has specified a prefix length. If not, or if
the prefix length equals the maximum prefix length for the address
family, we assume it to be a host and search using the contains
operator. This is identical to the old functionality.

If the user has specified a prefix length which is not the maximum
prefix length we instead coerce the prefix according to the prefix
length by zeroing out any bits set to the right of the mask and then
search with the equal operator, so if 10.0.0.1/24 is specified, we
search for 10.0.0.0/24, find the assignment (as per the above example)
and can set type to 'host'. For adding 10.0.0.0/8 (again per the above
example) we will do an equal search for 10.0.0.0/8, which won't return
anything and thus we will not try to override the specified type so the
prefix can be correctly created, or if no type is set we will return an
error.

Fixes #800.