Add unicode support to whoisd
The whois protocol specifies ASCII encoding of characters which means
there is no standardized way to support non-ASCII characters. IDN is
used for supporting non-ASCII characters for DNS and it appears that
multiple whois clients use IDN if non-ASCII characters are detected, so
we try to decode with IDN and if it fails we fail over to normal
decoding.

Responses are now sent in UTF-8 which isn't exactly kosher but it
appears other whois servers do the same (whois.nic.se for example) so I
suppose it's okay.

Fixes #808.