whoisd: assume utf-8 encoded queries
The popular whois client written by Marco d'Itri does IDNA coding for
queries when non-ASCII characters are detected, as does some other whois
cilents. However, there are corner cases where it does not. Typically
whois is used to lookup domain names so the query input is a single
word. NIPAP however supports multiple words and for an input such as "å
ä ö" the resulting query is "å ö xn--nda", which is a bit peculiar. The
first two letters are UTF-8 encoded while the last, "ö", is IDNA
encoded.

The whois server now uses the StreamRequestHandler, which is a more
elegant and Pythonic way of receiving queries since it works on whole
lines instead of bytes. Each line is then UTF-8 decoded and we try to
IDNA decode it, but this only works for single-word input.

This means that most cases do work, ie single-word non-ASCII queries are
IDNA encoded by the client and we successfully decode it. Multi-word
non-ASCII queries are sent as UTF-8, which we also decode, *unless* the
last character of the query is non-ASCII, in which case the whois client
mixes UTF-8 and IDNA. I'm not sure if this is intended or not, but we
don't handle it.

Part of #808.