www: Update VRF filter list on VRF removal
Make sure that the VRF filter list is updated when a VRF in the filter
list is removed from NIPAP. Previously when this happened, the web UI
ended up in a state where a VRF in the filter list could not be removed
from the filter list (as it couldn't be searched for).

Also somewhat handles this for other users that the one removing the VRF
by making sure the VRF filter list is up to date whenever the search box
is opened.

Fixes #811