Complete basic smart parser
Not sure I should call this completed or not but at least it works the
way I want for all queries I've tried so far. There are certainly areas
to improve still but it's getting there.

The smart parser now runs the test suite without failure. Only change to
the test suite is the text expected for a search string with an unclosed
quote.

For each object type (VRF, Pool, Prefix) we have a class that inherits
from the main SmartParser class and applies logic specific for the
object type so we end up with VrfSmartParser, PoolSmartParser and
PrefixSmartParser. For example, only PrefixSmartParser matches on
prefixes.

Exceptions thrown by the backend are moved to a separate file so it can
be imported both into backend.py and smart_parsing.py

Fixes #531.