smart-parse: improve parsing of expressions
The way we build the groupings for pyparsing is reflected in the
results. Setting the results name of a component that is part of another
component yields ParseResults of different "depth" where we have to go
through two lists to access the value we want. The real problem is that
things like vrf_rt was built out of different components and so the
result would come in at different depths compared to a 'word' or
something else. Handling different depths is a PITA so we avoid that by
not building components out of other named components.

The interpretation of "expressions" has been improved too, partly by
following the interpretation format used elsewhere and by including
relevant information.

Part of #531.