sql: fix parent_prefix selection
If a prefix is shrunk we have to filter it out in the before trigger to
avoid selecting its old self as its parent_prefix. This is perfectly
natural as the database before trigger sees the data before the update
has been applied.

This fixes the scenario when trying to shrink an assignment prefix.
Before it would select itself as parent and when running the consistency
checks on that it would refuse the update as an assignment cannot have
an assignment (its old self) as a parent.

The prefix statistics were written under this false assumption, so with
the correct parent_prefix it would generate incorrect statistics which
I've solved by adding an explicit path for when shrinking prefixes, i.e.
NEW is smaller than OLD but indent level is the same.

This fixes #817.