cli: Add config file option for custom columns
This allows the user to specify custom columns in their .nipaprc file in
addition to being able to specify it on the command line. We prefer the
command line and then look in config file.

I had made a mistake around parsing the shell_opts, assuming that the
--columns argument was always there and so the standard, ie when not
specifying custom columns wouldn't work.

Also widens the column for tags from 1 to 2 characters, which is need
when tags are set to keep alignment.

Part of #684.