cli: force UTF-8 output on UTF-8 terminals
This is really rather weird but raw_input has an issue with UTF-8 output
even when stdout encoding is already set to UTF-8. Forcing UTF-8 for
UTF-8 terminals, which really should be a no-op, fixes that issue.

Since we are checking for UTF-8 output, this should be compatible with
legacy terminals.

Redirected or piped output is forced to UTF-8 just as before.

Fixes #841.