Drop dependency on PostgreSQL
First off this allows installation on newer platforms which have
versions of PostgreSQL that we haven't specified in the control file.
One would think it's enough with 'postgresql > 9' but there is no such
meta package for ip4r, instead the versions are called
postgresql-ip4r-9.1 or similar so we have to list them individually.

Secondly, not everyone wishes to run Postgres on the same machine as
nipapd is running and that wasn't possible before.

This does add a step to the installation procedure and it's not ideal
that the postinstall script tries to fiddle with the database before we
know it's actually there but I intend to deal with that in #715.

Fixes #894.