backend: Avoid setting extra query to empty dict
If the extra_query attribute was left out of a smart_search_pool-query to the
XML-RPC-backend, an empty dict was sent to the internal backend function
instead of the intended 'None' object. As an empty dict isn't a proper
dict-to-SQL-dict (lacks operator, val1 and val2) the query failed (quite ugly -
a standard python KeyError exception was returned).

This commit adresses that issue by simply passing the None object along to the
internal backend function.

Fixes #910