www: Update AngularJS and UI-Bootstrap
Updated AngularJS to version 1.5.3 and UI-Bootstrap to version 1.3.3.

Main reason for upgrade is to get versions which are compatible
with each other. Last time AngularJS was updated UI-Bootstrap was
not which broke the tooltips.

Also cleaned up slightly in the mess of javascript libraries; added
version numbers and removed non-minified versions of the libraries.