www: Disabled Angulars Strict Contextual Escaping
Disables the Strict Contectual Escaping (SCE) feature in AngularJS.

This is a pretty bad thing as the SCE
(https://docs.angularjs.org/api/ng/service/$sce) assists in writing
secure code invulnerable to XSS attacks. The reason for doing this is
that I cannot get the tags tooltip to work in the prefix list, which is
currently not implemented in AngularJS and where I did not manage to
reach $sce.trustAsHtml function necessary for allowing HTML in tooltips.
A pretty ugly hack was also needed to get the tags tooltip to work on
the AngularJS-based VRF and pool list pages; to avoid
"[$rootScope:infdig] 10 $digest() iterations reached. Aborting!" the
result after passing the HTML through $sce.trustAsHtml function needed
to be stored globally for each tags tooltip that was generated.

The implication of this is that an attacker can insert nodes into the
DOM by adding a specifically crafted tag to a VRF, pool or prefix.

This should be re-evaluated when the prefix list is implemented in
AngularJS.